new_variant: update location of ec.bin
The location for the EC binary changed; it now lives in
ec/build/${VARIANT}/ec.bin. Update new_variant.py to check this
location to determine if the EC build was successful.

BUG=b:152350312
TEST=Run new_variant.py to create a new variant of the hatch
baseboard. Generate the fitimage outside the chroot as the program
instructs. Resume new_variant.py. Observe that the ec_image step is
able to build the EC image for the variant and find ec.bin when it's
done, then continue on to ec_buildall.

Change-Id: I6c2841d0efae95e59577c99893393ca0c8d4831f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2120261
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jacob Rasmussen <jacobraz@google.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>