new_variant: only check config.yaml for variant name
Due to changes in other CLs, the emerge_yaml step does not create
a config.json file now, and will not create config.c in the very
near future. The only one of these files that will continue is
config.yaml, so just look for the variant name in that file.

BUG=b:152350811
TEST=Run new_variant.py to create a new variant of the hatch
baseboard. Generate the fitimage outside the chroot as the program
instructs. Resume new_variant.py. Observe that the build_yaml
step is succesful.

Change-Id: Id8a28b8d8a10ad725624ea242d7f2b628d04e3ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2120262
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>