new_variant: add Cq-Depend to uploaded CLs
Add a dependency graph for each of the reference boards. After all
of the CLs are uploaded (and the coreboot CL has been upstreamed,
if applicable), walk the dependency graph and assemble a Cq-Depend
line for each CL that has dependencies. Amend the commits that need
to have Cq-Depend added, and re-upload those CLs.

BUG=b:151731712
TEST=Test the logic that assembles the Cq-Depend string by following
instructions in testdata/README.md, and see that the program prints
out the correct Cq-Depend strings for the CLs that need them.

Start a new hatch variant named 'sushi' (inside the chroot).
(cr) $ ./new_variant.py --board=hatch --variant=sushi
Generate the fitimage outside the chroot when asked.
Resume the program (`./new_variant.py --continue`)
Wait for the EC and then the chromeos-bootimage to all build.
Push the coreboot CL to review.coreboot.org when asked.
Resume the program. CLs get uploaded, then the program asks to wait
for the coreboot CL to get upstreamed. Get someone to review and
approve, then upstream.
Resume the program. It determines the Cq-Depend, amends the
commits, and re-uploads the changed commits to gerrit.
Check all of the commits that have been uploaded to publid and
private gerrit, and ensure that the Cq-Depend lines are correct.

Cq-Depend: chromium:2125164, chromium:2125163, chromium:2125270, chrome-internal:2820196, chrome-internal:2819185
Change-Id: I4a0a4b49308cef86b2834170784440b7c517ec15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2129197
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>