devserver: Deprecated hostlog API
This API was being used when we were starting a new devserver instance
along with lab devservers, but we don't do that anymore. Furthermore, we
have moved to update_engine.log based hostlog (crrev.com/c/2120287) so
this can be removed. The lab devservers don't even turn on this flag.

BUG=chromium:1064135, chromium:872441
TEST=devserver_integration_test.py

Change-Id: I69f0faeae2fe11638f7898f5ff4674ca00c218b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2121395
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Sanika Kulkarni <sanikak@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>