new_variant: replace file_exists() with isfile()
os.path.isfile() checks that the file exists and that it's a file.
Since we only need to call one function, just remove file_exists()
and have the caller join the path and filename before calling
isfile().

BUG=None
TEST=Run new_variant.py to create a new variant of the hatch
baseboard. Generate the fitimage outside the chroot as the program
instructs. Resume new_variant.py. Observe that the steps are
successful, and the program stops to ask the user to push a CL
to review.coreboot.org.

Change-Id: Ic178d158fe6ec3ea17219c494d618ad66c3fcc5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2121364
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>