create_initial_ec_image: rename variable and reword messages
Change the name of the BASE variable to REF. Change error messages
and the commit message to use the term "reference board" instead of
"baseboard". Note that new_variant.py was already passing the name
of the reference board to this script; we were just using a variable
name in the script that implied it was the baseboard, and calling it
a baseboard in error messages and the commit message.

BUG=b:149334328
TEST=`create_initial_ec_image.sh hatch tiamat`
Check the commit in platform/ec and verify that it refers to hatch
as the reference board.
Run the script with invalid reference board names, and verify that
the error messages use the term "reference board."

Change-Id: Ibd290fbdd15ccab178c6537dcc00d206d81ea82a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2151726
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>