findmissing/synchronize: Reduce noise when running tool commands
We don't want to display log commands while running tool commands.
At the same time, we do want to see activity logs if the synchronization
script runs on GCE. To address the problem, introduce logging facility
and start using it for database synchronization. Set the log level for
local commands to WARNING and for commands executed on gce to INFO.

BUG=None
TEST=Run scripts on target

Change-Id: Ia15cbde84bc810e8340234930565158acfb81ebd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2189517
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>