findmissing: fix: Reset HEAD back to origin/branch
Issue: reset_and_clean_kernel did not reset local branch commits to
origin. This caused commits to build ontop of the last and create a
chain of CL's.

Fix: checkout_and_clean now resets the current branch of the kernel
path to be in sync with it's respective origin/branch. This means that
when we call this method, our local branch will be match it's origin/branch.

BUG=None
TEST=None

Change-Id: I9f20e1fb8f290c8b358df6f73992a5855dd80d1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2154114
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>