findmissing: Clone stable release candidate repository when running locally
We need the stable release candidate repository to be able to check
if a patch is queued there for a future stable release. This is only
needed if local commands are executed, so add 'local' argument to
synchronize_repositories() and only synchronize the release candidate
repository when running an explicit command (ie not as part of the scripts
running on GCE).

BUG=None
TEST=Run getopen command

Change-Id: I1ca4e4b13b61da3484c67badea50b3b5c8fcf235
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2200285
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>