findmissing: Modify sync scripts to update by checksum
Issue with syncing files if destination timestamp was updated.
When syncing from local -> gce, or findmissing repo -> production code,

the previous rsync option to update (-u) is documented as the following.
-u  skip files that are newer on the receiver

This caused issues syncing if there was changes made locally on the
destination directory.

Modified to use checksum option which should transfer files from source
to destination if the files are not the same.

-c  skip based on checksum, not mod-time & size

BUG=None
TEST=None

Change-Id: I954b59a205852ad9b947a80c5159d0f3030c70fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2163187
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Hirthanan Subenderan <hirthanan@google.com>
Tested-by: Hirthanan Subenderan <hirthanan@google.com>