new_variant: don't look for fit.log
Remove the check for fit.log as part of verifying that new_variant
can continue with the build. The name of the output log file can
differ, depending on the baseboard, so log file is not always named
fit.log. It is sufficient to check for the fitimage and the
-versions.txt file.

BUG=b:151472333
TEST=start a new variant of the volteer baseboard. Generate the
fitimage outside the chroot as instructed. Continue new_variant.py
and see that it detects the fitimage binary and -versions.txt file,
but does not error out on a missing fit.log (because the log file
is named fit-${VARIANT}.log for volteer).

Change-Id: Ic7939ff49fc61b49574796d9aa3da3461993a90c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2171704
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Nick Vaccaro <nvaccaro@chromium.org>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>