new_variant: add project_config step for volteer
For projects that use the new starlark-based configuration structure
instead of model.yaml, we need a gerrit admin to create the new
project and populate it. new_variant.py will verify that emerging the
private BSP config will result in a project-config.json that contains
the name of the new variant, and then we can continue creating the
other CLs for the new variant's firmware.

BUG=b:154115910
TEST=Create a new variant of Volteer when the project config repo does
not exist. You get an error message that you need a gerrit admin to
create the repo first.

Create a "fake" repo for the project config. See
* https://buganizer.corp.google.com/issues/154115910#comment7
* https://buganizer.corp.google.com/issues/154115910#comment16
for details about how to create it in the right directories.
Run new_variant.py again, and it will fail, because the new
project isn't included in the ebuild.

Add the new variant's name to PROJECTS in
src.private-overlays/overlay-volteer-private/chromeos-base/
 chromeos-config-bsp-volteer-private/
 chromeos-config-bsp-volteer-private-9999.ebuild
Run new_variant.py again, and it will successfully build
project-config.json, then move on to the coreboot variant, coreboot
config, and fitimage. Once it asks you to generate the fitimage,
use --abort to clean up.

Change-Id: I5dfe4c32f9fbacac788f7ba97ae5e068fbf097e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2171706
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>