findmissing: gerrit_interface: Fall back to use emails from tags if needed
In create_change(), we try to assign reviewers based on the Gerrit
Change-Id of the commit being fixed. In some situations (for example, if
the commit was pushed into Gerrit through a merge) the commit may
include a Change-Id, but that Change-Id is not known to Gerrit.
Pick reviewers from tags found after "cherry picked from" in this
situation.

BUG=None
TEST=Run on GCE

Change-Id: Ibe8994676ba85a8f562ae3fc74d0c5997a3b6eb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2197186
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>