quick-provision: Don't do wget retries
It seems like wget internally doing retries won't work if the server is
immediately not recovered which probably is most of our use cases in the
lab. This is causing the gzip not work properly when wget retries and
gets the payload from the begining. This CL turns of the wget retires.

BUG=chromium:1068388
TEST=None

Change-Id: Id005575c3b1e4dc0f614fc11725ba005d060f5a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2199483
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>