findmissing: initdb_chromeos: Ignore duplicates when updating commit table
If a merge was handled previously and the last commit handled was part of the
merge, the next run of the script will report lots of duplicates if the merge
was non-linear. This generates a lot of confusing error messages. Let's ignore
this situation.

BUG=None
TEST=Run on GCE

Change-Id: I67cbc2aa1820f18dfce868ecaaf40dbf435533db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2207394
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>