findmissing: Find subsequent patches to fixes
It looks a bit silly if we are constantly submitting fixes for fixes if
know the subsequent patches before hand. Solution is to do an interative
lookup.

Note this solution is temporarily and will be replaced by
crrev.com/c/2208004 once the database supports CTEs.

BUG=None
TEST=Pass in known test SHA

Change-Id: Ia334b028378c9a8aa1aaebaf175369bc65dd73d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2146069
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>