cvetriager: patch applier
The patch applier is designed to collect a commit from the LINUX kernel
and attempt to cherry-pick it to the Chrome OS kernels. If
successful, then the patch is formatted to contain important
information, such as bug id, cherry-pick source tag, and a sign-off tag.

BUG=chromium:1093363
TEST=python setup.py test

Change-Id: I2f6139d2414f78e38294c0dc234a70583e9db0d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2230699
Reviewed-by: Zubin Mithra <zsm@chromium.org>
Commit-Queue: Wanda Mora <morawand@chromium.org>
Tested-by: Wanda Mora <morawand@chromium.org>