contrib: utility script to assist with unibuild conversions
Utility script to assist with converting pre-unibuild boards to
unibuild v2 (model.yaml format).

Run with --help for usage, you'll need a DUT over SSH that the script
can run commands on.

Note: this is a contrib-quality (unsupported) script, and it's meant
to be this way.  Essentially just submitting as is for now so that we
can iterate on it as we find more corner cases.

BUG=none
TEST=run for nocturne, build image and flash

Change-Id: I56abc90db0df6c8fe091add5181e8226f08a5b95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2233191
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Sean McAllister <smcallis@google.com>