new_variant: enable default firmware build
Add a script to set _FW_BUILD_CONFIG in config.star, regenerate the
config, and commit the results. Have all of the reference boards that
use per-project configuration repos add a step to call this new
script. The resulting CL has to Cq-Depend on all of the other CLs
that create the firmware.

BUG=b:158602858
TEST=`./new_variant.py --board=puff --variant=wyvern`
Verify that a new CL is staged in src/project/puff/wyvern with
changes to config.star and the generated files.

Change-Id: I5ad4a6dcc079890a402b78a20e38c36b16d04711
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2238485
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>