contrib/dut-console: Don't use dut_control making us more robust
There's no real reason to use dut-control to talk to servo.  We can
use xmlrpc directly and it's super simple.

This allows us to get rid of the hacky include of python stuff in the
chroot which breaks as the chroot stops using python2.

NOTE: just to be clean, we'll also run xmlrpm with python3.  If this
causes problems for people we can try to come up with stuff that's
compatible with python2 too.

BUG=None
TEST=Open some consoles.

Change-Id: I8e58351567aa1f528361a45e6678d93c46c932ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2252320
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>