cvetriager: patchapplier: create new branch to cherry-pick onto
Update on patchapplier to create and checkout to a new branch
"b{bug-id}-{kernel}" to perform the cherry-pick onto. Also includes
small changes to variables used in the test file, and moving get_sha()
from patchapplier to its test file because it is only used in there.

BUG=chromium:1093363
TEST=python setup.py test

Change-Id: Ib206dfdd700553174dc7a3a29688c7d84875823c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2260574
Reviewed-by: Zubin Mithra <zsm@chromium.org>
Commit-Queue: Wanda Mora <morawand@chromium.org>
Tested-by: Wanda Mora <morawand@chromium.org>