new_variant: remove cb_variant from cq-depend
The cq-depend line can only specify CLs on the chromium or
chrome-internal gerrit instances. The 'cb_variant' step is a CL in
the coreboot gerrit, so remove it from the depends list for all
variants that use coreboot.org.
For Zork variants that are using coreboot-zork, remove the 'find'
step, as this step looks for an upstream CL that we won't have on
Zork for now.
To ensure that `git filter-branch` (which adds the Cq-Depend line)
will succeed, add the -f option so that remnants of previous
filter-branch operations won't cause a failure.

BUG=b:159812946
TEST=create a new variant and get it to the add_cq_depend step. Verify
that the fw_build_config CL does not include a CL in its cq-depend
line that starts with coreboot:

Change-Id: Ia59d69df40413ad4bc288f9bd5524641996ee7f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2264519
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>