cros_unibuild_convert: add wallpaper property
Detect if there is a pre-unibuild custom wallpaper.  This replicates
the logic in chromeos-assets-9999.ebuild: we search the wallpaper
source directory to see if there is a wallpaper under the board
name (or model, as some older boards have the variant naming) and use
that.

Under private overlay, as the wallpapers are in chromeos_interal
repo.  No use handing this out in public overlay.

BUG=chromium:1098942
TEST=generated unibuild config for atlas, wallpaper set to "atlas"

Change-Id: If7e242f11966a0e926d44971507cbd93ca1024f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2265260
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Sean McAllister <smcallis@google.com>
Commit-Queue: Sean McAllister <smcallis@google.com>