emacs: xml parsing rework.
Parsing XML within Emacs was too slow (over a minute). So we added a
utility Python script which parses the repo manifest faster. During
this rework, we also discovered that we need dest-branch in addition
to project name to identify a project path in a repo manifest.

BUG=chromium:1102923
TEST=Evaluated and called functions in IELM

Change-Id: Iecdb65d868e39b9d4695c31230f5d137f2e6e383
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2295702
Tested-by: Aaron Massey <aaronmassey@chromium.org>
Commit-Queue: Aaron Massey <aaronmassey@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>