cvetriager: new flag for detecting commits
Added a new flag to contextgenerator.py constructor to determine whether
detect_relevant_commits should be ran. The flag will be turned on by the
bin/ script if the user chooses to do so. This patch also includes the
deletion of an unused function in contextgenerator.py that went
unnoticed.

BUG=chromium:1093363
TEST=python setup.py test

Change-Id: I5cd338849b2f203e77e3e0ba476676a7021ffd49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2300006
Reviewed-by: Zubin Mithra <zsm@chromium.org>
Commit-Queue: Wanda Mora <morawand@chromium.org>
Tested-by: Wanda Mora <morawand@chromium.org>