findmissing: Add 'status' command
Sometimes it is useful to search the status of upstream commits
in the internal databases. We already report this status as part of
'abandon' and 'restore' commands if more than one entry is present.
Use that code to implement a new 'status' commands to report
the status of a given upstream SHA.

BUG=None
TEST=Run "findmissing status" command

Change-Id: I38878924e63add82f6e982b5ee1f2d5a93ef670b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2308253
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>