cvetriager: new json argument in bin/triage
A new argument is added to the bin/ script that prints out a JSON string
containing the results of each branch after attempting to patch a
commit. This will be used for informational purposes, for when a patch
fails to apply because it is already waiting to be merged in the working
tree.

BUG=chromium:1093363
TEST=python setup.py test

Change-Id: I5d03b2000178e25b7b241076d170e49c95494e66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2308750
Reviewed-by: Zubin Mithra <zsm@chromium.org>
Commit-Queue: Wanda Mora <morawand@chromium.org>
Tested-by: Wanda Mora <morawand@chromium.org>