update_program_fw: Add version validation
Add validation to check for major_version and minor_version
in files that require them.

The validation is not comprehensive since the configs allow versions to
be inherited from the base program.star, and projects can be skipped.

An explanatory message is provided to explain when a version string
should be added to a file.

BUG=none
TEST=Dry runs with zork and puff.

Change-Id: Id75e9f8549cda7833fef49c25939a419797f8c16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2315777
Commit-Queue: Andrew McRae <amcrae@chromium.org>
Tested-by: Andrew McRae <amcrae@chromium.org>
Reviewed-by: Kangheui Won <khwon@chromium.org>