cvetriager: modify change id option for testing
A new test mode for the script where the change ids are modified to
point to test CLs instead of generating new ones. This is to avoid
generating too many CLs that need to be abandoned when testing the
script.

get_sha in patchapplier_test.py in now moved to common.py

BUG=chromium:1093363
TEST=python setup.py test

Change-Id: I72b0576fca58b8c1483acbf63a51de0c4f349c0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2317548
Reviewed-by: Zubin Mithra <zsm@chromium.org>
Commit-Queue: Wanda Mora <morawand@chromium.org>
Tested-by: Wanda Mora <morawand@chromium.org>