findmissing: cloudsql_interface: Fix list of SHAs in SQL query
Unfortunately, a single element in a tuple creates unwanted output in
python.

>>> l=['a']
>>> tuple(l)
('a',)

That means we can not use str(tuple(l)) to create a list of SHAs
for the SQL query since SQL doesn't like the trailing ','.
Go back to the clumsy method using .join().

BUG=None
TEST=Run findmissing with single SHA as argument

Change-Id: I26c80e924bf98d6f08f52da2704133bef396fa86
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2321780
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>