findmissing: Restore 100-character line length limit for python scripts
Commit a52c5a6bb88e ("pylintrc: set max line length") set a new line length
limit of 80 (from the old 100) for python scripts. The Gods Have Spoken,
but it doesn't make sense to have to rewrite existing scripts just to
follow new restrictions. Set it back to 100 for findmissing.

The pylintrc file is a copy of the parent directory's pylintrc with
the line limit changed from 80 to 100.

BUG=None
TEST=repo upload

Change-Id: Id3a4bc06289d4295a6fcc4a5dac000b56fadf78b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2352412
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>