new_variant: allow building on WIP
When making changes to the files that new variants are based on,
like the coreboot template files, or the EC baseboard source code,
we want to be able to base the CLs for the new variants on top of
the CLs we're working on, not on m/master. Add the NEW_VARIANT_WIP
variable to signal to the scripts that `repo start` should add
the `--head` option to base the new branch off the current HEAD,
and not on m/master.

BUG=b:167305316
TEST=create a CL in src/platform/ec with a simple change to a file
like some docs. Run `create_initial_ec_image.sh waddledoo kingitchy`
and then check that the other change is missing; the new branch is
based on m/master, not HEAD. `repo abandon` the branch that
`create_initial_ec_image.sh` created. Run it again with
`NEW_VARIANT_WIP=1 create_initial_ec_image.sh waddledoo kingitchy`
and then check that the other change is included. `repo abandon`
the new branch, and see that you've been reset back to m/master
and will need to `git checkout` the experimental branch (which
you should also `repo abandon`.

Change-Id: I5998179cbaefce75ed1e7d168b656ccf66c2f513
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2424754
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>