new_variant: updates for dedede boards
Dedede boards are using a 16MB SPI ROM now, and waddledoo-based
boards in particular are no longer using static ACPI for the
camera, so update the options written to Kconfig.name.
Because of the 16MB SPI ROM, we also need to use a different
fitimage in the end-to-end test so that the image will link
correctly.

BUG=b:168755528
TEST=`testdata/new_variant_fulltest.sh waddledoo` succeeds.
All other reference boards (waddledee, puff, hatch, trembyle, and
dalboz) also succeed.

Change-Id: I15d4455da2b9d7e0899f9485c1e092c661655245
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2425243
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>