gs_archive_server: add bind option to the gs_archive_server command
Currently the service by default runs on 127.0.0.1 when started. When
the service is containerized, it needs to be run on 0.0.0.0 so that
the service is accessible from outside of the container (host). Add a
flag '-b'/'bind' so that a different address can be supplied to
the service via the command line and then update the cherrypy global
configs accordingly.

BUG=chromium:1084807
TEST=Start the service by passing the -b flag and verify if it runs on
the correct address.

Change-Id: I359156c58f059c70e2e3190fd4d1f01a23ddc0fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2477498
Tested-by: Sanika Kulkarni <sanikak@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>
Commit-Queue: Sanika Kulkarni <sanikak@chromium.org>
Auto-Submit: Sanika Kulkarni <sanikak@chromium.org>